+++ SalishSeaTools/CHANGELOG.rst
@@ -4,6 +4,12 @@ Changelog
 Next Release
 ------------
 
+* ``hg_commands`` module and documentation:
+  A library of Python functions for working with Mercurial
+  via its command line interface.
+  This is a utility library that is used by the ``SalishSeaCmd` package
+  and other modules in the ``SalishSeaTools`` package.
+
 * ``nc_tools`` module and documentation:
   A collection of functions for exploring and managing the attributes of
   netCDF files.